+++ docs/changelog.md
@@ -42,6 +42,10 @@ pageflow_next_text: Contributing
   - added empty `Entity` interface
 * Added immutable support to entities
   - added `ReadOnlyForeverException` exception
+  - added `setReadOnlyForever()` to `WriteProtectableEntity` interface
+  - `WriteProtectableEntity::setReadWrite()` must now throw `ReadOnlyForeverException` if `setReadOnlyForever()` has been called
+  - added `setReadOnlyForever()` to `WriteProtectTab` trait
+  - `WriteProtectTab::setReadWrite()` now throws `ReadOnlyForeverException` if `WriteProtectTab::setReadOnlyForever()` has been called
 
 ### Refactor